Update README.md with APR references
Hello, the Arab Canadian Lawyers Association have created the Anti-Palestinian Racism Naming, Framing and Manifestations and brought it forward to Canada to adopt. Currently, it’s adopted. This working definition allows Palestinians and their supporters to advocate for the adoption of this Framework. There is a bullet point summary in the fact sheet by the CJPME that sums up the forms of Racism as well. We need academic & healthcare institutions, government organizations and corporations to adopt the framework so people can Name it and Report it when the racism occurs. Thank you!!